fix[lang]: transitive exports (#3888)
this commit fixes transitive exports, i.e. where `module1` exports
`module2.foo`, but it was previously not exportable from `module1`. this
is because previously exported functions were not added to the
`ModuleT`'s `exposed_functions` list.

a test had previously been written for this case, but it had been marked
xfail since the desired behavior was not clear until user feedback was
received. this commit removes the `xfail` marker from that test case.